refactor(a1): extract /api/events search to routers/catalog.py (BR-CODE-1 slice 6)
Moves the /api/events TEvo search passthrough out of app.py into the existing
routers/catalog.py factory (get_client + require_auth). +1 test pinning the
arg + popularity-order_by passthrough. Behavior identical.
