+++ KANBAN.md
@@ -48,7 +48,7 @@ Prioritized program from the 2026-06-19 cross-axis bottleneck audit (code · dat
 |---|---|---|---|---|
 | **P0** | → `A1-OPS-3` | A1 / Op | **SG broker token is the #1 product ceiling** — one token shared with an invisible external prod program; 63–77% 429, 1,235 starved events, 10/30 MLB games 2–3d stale. Physics limit, not code. **⏸ SG token-consumer crons PAUSED 2026-06-19 (operator-directed): 6 jobs `active=false` (355/248/64/63/381/407) — see `docs/archive/2026-06-19-snapshot-bloat-play.md` for the list + resume SQL.** | See `A1-OPS-3` (OPS table). **Operator action:** coordinate SG token isolation from the external app; then resume the 6 paused crons. |
 | **P0** | → `A1-OPS-16..21` | A1 / Op | **Firehose analysis SPEED** (operator reframe 2026-06-19: speed, not disk — disk is cheap). `listings_snapshots` 48GB/123M with **22GB bloated indexes** (415M deletes churn them); `min(captured_at)` **times out at 25s**. Pre-agg layer is ~200× smaller (`latest_event_metrics` 20MB). | **Speed play: `docs/archive/2026-06-19-snapshot-bloat-play.md` §Speed solution** — 4 pillars: (1) route analysis to pre-agg layer (SOP) → (2) **drop dead indexes — migration AUTHORED `20260619200000_drop_dead_firehose_indexes.sql`, operator-apply** → (3) `REINDEX CONCURRENTLY` hot indexes → (4) partition by `captured_at` (pruning + DROP-PARTITION retention = no re-bloat). Plus espn `pg_repack`+autovac tune. Apply = A1/operator (ship-a-migration). |
-| **P1** | **BR-CODE-1** | D0 / A1 / Op [IN PROGRESS] | **`app.py` monolith — 8,753 LOC, 112 routes, 236 fns, 302 inline SQL calls, ALL surfaces in one file** (5 fns >220 lines). #1 velocity drag; every route re-writes the same filter chains, no DB-access layer. | Carve out a `routers/` package by domain (broker · store · seatgeek · seatdata · axs · admin) + a thin shared `db`/query helper; move routes incrementally behind unchanged paths. Land in small PRs, tests green each step. **Slice 1 (done):** `routers/` package + `routers/site_essentials.py` (robots/sitemap/favicon, factory takes deploy base URL). **Slice 2 (done):** `routers/catalog.py` — TEvo performer/venue/configuration detail passthroughs (live `client` getter + `require_auth`). **Slice 3 (done):** `routers/lists.py` — `/api/watchlist` GET + `/api/runs` + `/api/snapshots/latest|velocity` (live `require_sb` getter + `require_auth`; watchlist POST/DELETE stay in app.py; covered by `test_watchlist_runs_snapshots.py`). **Slice 4 (done):** `core/config.py` — env-derived config out of app.py bootstrap. **Slice 5 (done):** `core/auth.py` — the security-critical **`require_auth` gate + `AUTH_DISABLED` kill-switch + `_is_production`** moved out of app.py (verbatim); app.py re-exports them (`app.require_auth is core.auth.require_auth`); only the security test's `AUTH_DISABLED` monkeypatch repointed to `core.auth` (the `requests` stub hits the shared module, unchanged). Security suite 6/6 + 122 route net = 128 green. **Next slices:** move `sb`/`client` (+ creds bootstrap) into `core/` → `routers/public_meta.py` (`/api/public/config`+`/version.json`+`/healthz`) → then the big broker/store route blocks (need shared helpers in `core/`). |
+| **P1** | **BR-CODE-1** | D0 / A1 / Op [IN PROGRESS] | **`app.py` monolith — 8,753 LOC, 112 routes, 236 fns, 302 inline SQL calls, ALL surfaces in one file** (5 fns >220 lines). #1 velocity drag; every route re-writes the same filter chains, no DB-access layer. | Carve out a `routers/` package by domain (broker · store · seatgeek · seatdata · axs · admin) + a thin shared `db`/query helper; move routes incrementally behind unchanged paths. Land in small PRs, tests green each step. **Slice 1 (done):** `routers/` package + `routers/site_essentials.py` (robots/sitemap/favicon, factory takes deploy base URL). **Slice 2 (done):** `routers/catalog.py` — TEvo performer/venue/configuration detail passthroughs (live `client` getter + `require_auth`). **Slice 3 (done):** `routers/lists.py` — `/api/watchlist` GET + `/api/runs` + `/api/snapshots/latest|velocity` (live `require_sb` getter + `require_auth`; watchlist POST/DELETE stay in app.py; covered by `test_watchlist_runs_snapshots.py`). **Slice 4 (done):** `core/config.py` — env-derived config out of app.py bootstrap. **Slice 5 (done):** `core/auth.py` — the security-critical **`require_auth` gate + `AUTH_DISABLED` kill-switch + `_is_production`** moved out of app.py (verbatim); app.py re-exports them (`app.require_auth is core.auth.require_auth`); only the security test's `AUTH_DISABLED` monkeypatch repointed to `core.auth` (the `requests` stub hits the shared module, unchanged). Security suite 6/6 + 122 route net = 128 green. **Slice 6 (done):** `/api/events` search → `routers/catalog.py` (extends the existing factory; +test). **Next (larger, planned):** the `sb`/`client`/`require_sb` → `core/` move is HIGH-blast-radius (creds bootstrap has `sys.exit`; monkeypatched across many test files) — plan it deliberately, not a quick slice. Then `routers/public_meta.py`, then the big broker/store blocks (need shared helpers in `core/` first). |
 | **P1** | **BR-CODE-2** | A1 / D0 | **9 `*_client.py` copy-paste ~100 LOC each** (read-only guard, retry/backoff, vault-secret, session/headers) with *inconsistent* retry → uneven rate-limit handling (ties into the SG 429 story). | Extract a `BaseReadOnlyClient` (GET-only guard + unified exponential backoff honoring `ratelimit-remaining` + vault resolver); migrate clients one at a time. ~900 LOC reclaimed; `tests/test_readonly_guards.py` stays green. |
 | **P2** | **BR-CODE-3** | B1 / D0 [IN PROGRESS, PR #584] | **Test coverage ~12%** (5,951 test LOC vs ~46.5k); **60+ broker/admin routes untested** (`/api/broker/*/chart-data`, `/movers`, all `/api/admin/*`, `/api/seatdata/*`, `/api/axs/*`). | Add route-level tests for the broker/admin surface (mock the SECDEF RPCs); prioritize the event-page + movers paths. Target 30% ratio. **Increment 1 (done):** `tests/test_public_infra_routes.py` — 10 tests over the public non-DB infra routes. **Increment 2 (done):** `tests/test_broker_routes.py` — 5 tests over `/api/broker/leagues`, `/performer/{id}/assets`, `/event/{id}/overview` + a reusable `FakeSupabase`. **Increment 3 (done):** `tests/test_broker_movers.py` — 4 tests over `/api/broker/movers` v2 + `/watchlist-movers`. **Increment 4 (done):** `tests/test_watchlist_runs_snapshots.py` — 6 tests over `/api/watchlist`, `/api/runs`, `/api/snapshots/latest` (event-join enrichment + sort), `/api/snapshots/velocity`. **Increment 5 (done):** `tests/test_catalog_routes.py` — 6 tests over the TEvo-catalog passthrough GETs (`/api/performers/{id}`, `/api/venues/{id}`, `/api/configurations` + `/{id}`) asserting arg-passthrough + response relay. **Total: 31 route tests.** **Next:** `/event/{id}/chart-data` (469-line fn) + the `/api/admin/*` POST orchestrators (mock the upstream fire). |
 | **P2** | → `A1-OPS-22` | A1 / Op | **AQ map duplicates/false-matches** — 235 `tevo_event_id`s with ≥2 rows → cross-source visibility gaps; safe-class consolidation fn built, pending apply. | See `A1-OPS-22`. Operator apply + dry-run review → execute. |
