+++ routers/catalog.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 
 from typing import Callable
 
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, Query
 
 
 def build_catalog_router(get_client: Callable[[], object], require_auth: Callable) -> APIRouter:
@@ -38,5 +38,26 @@ def build_catalog_router(get_client: Callable[[], object], require_auth: Callabl
     @router.get("/api/configurations/{config_id}")
     def configuration_detail(config_id: int, _=Depends(require_auth)):
         return get_client().get_configuration(config_id)
+
+    @router.get("/api/events")
+    def events_search(
+        q: str | None = None,
+        performer_id: int | None = None,
+        venue_id: int | None = None,
+        occurs_at_gte: str | None = Query(None, alias="occurs_at.gte"),
+        occurs_at_lte: str | None = Query(None, alias="occurs_at.lte"),
+        only_with_available_tickets: bool = True,
+        _=Depends(require_auth),
+    ):
+        events = get_client().search_events_all(
+            q=q or None,
+            performer_id=performer_id,
+            venue_id=venue_id,
+            occurs_at_gte=occurs_at_gte,
+            occurs_at_lte=occurs_at_lte,
+            only_with_available_tickets=only_with_available_tickets,
+            order_by="events.popularity_score DESC",
+        )
+        return {"count": len(events), "events": events}
 
     return router
